requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
shap
dash
plotly
PyJWT
bcrypt
kafka-python
mysql-connector-python
yfinance
python-dotenv
sqlalchemy
dash-bootstrap-components
requests
apscheduler
confluent-kafka
python-jose[cryptography]